moved fixtures to conftest.py
