+++ tests/utils/database.py
@@ -1,7 +1,6 @@
 from application.db.models import EntityOrm, DatasetOrm, TypologyOrm
 from application.db.session import get_session
 from tests.test_data import datasets, entities
-import pytest
 
 
 def add_entity_to_database(entity: EntityOrm):
@@ -17,35 +16,6 @@ def add_entities_to_database(entities: list[EntityOrm]):
         session.add(EntityOrm(**entity))
     session.commit()
     return entities
-
-
-# this function ensures that a database has only the base elements
-#   - resets the database
-#   - adds the entities and datasets from tests/test_data/datasets.csv and tests/test_data/entities.csv to the database
-#   - yields back to the test
-#   - resets the database
-@pytest.fixture()
-def add_base_entities_to_database_yield_reset():
-    reset_database()
-    add_base_entities_to_database()
-    add_base_datasets_to_database()
-    add_base_typology_to_database()
-    yield
-    reset_database()
-
-
-# this function ensures that a database is empty for a test by...
-#   - resets the database
-#   - adds the base datasets and typology to the database
-#   - yields back to the test
-#   - resets the database
-@pytest.fixture()
-def empty_database():
-    reset_database()
-    add_base_datasets_to_database()
-    add_base_typology_to_database()
-    yield
-    reset_database()
 
 
 # this function adds the base entities to the database
